requirements.txt
albumentations>=1.3.0
argparse==1.4.0
datasets==4.4.1
h5py>=3.8.0
ipython>=8.12.0
jupyter>=1.0.0
lightning==2.5.6
transformers==4.57.1
torchvision==0.24.1
matplotlib>=3.7.0
numpy>=1.24.0
openai==1.59.5
opencv-python>=4.8.0
pandas>=2.0.0
peft==0.18.0
pillow==12.0.0
python-dotenv==1.0.1
pyyaml>=6.0
scipy>=1.10.0
seaborn>=0.12.0
tensorboard>=2.13.0
torch==2.9.1
torchvision==3.0.3
tqdm>=4.65.0
transformers==4.57.1
ultralytics>=8.0.0
wandb>=0.15.0
